separate the activity entry from original screen
