+++ constants/Spacings.js
@@ -2,6 +2,7 @@ const Spacings = {
   // margins:
   narrowMargin: "2%",
   mediumMargin: "5%",
+  wideMargin: "7%",
 
   // gaps:
   narrowGap: 10,
