+++ screens/DietEntryScreen.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import DietEntry from '../components/DietEntry';
+
+const DietEntryScreen = () => {
+
+  return (
+    <DietEntry />
+  )
+}
+
+export default DietEntryScreen
